+++ Biovation.Service.Api/v2/UserService.cs
@@ -43,7 +43,7 @@ namespace Biovation.Service.Api.v2
             return await _userRepository.DeleteUser(id, token);
         }
 
-        public async Task<ResultViewModel> DeleteUsers(List<int> ids = default, string token = default)
+        public async Task<ResultViewModel> DeleteUsers(List<long> ids = default, string token = default)
         {
             return await _userRepository.DeleteUsers(ids, token);
         }
